requirements.txt
aiohttp==3.6.2
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.3.4
entrypoints==0.3
idna==2.8
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pokebase==1.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2019.3
pyzmq==19.0.0
requests==2.22.0
Send2Trash==1.5.0
six==1.14.0
spotipy==2.12.0
terminado==0.8.3
testpath==0.4.4
tk==0.1.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
yarl==1.4.2
